友達登録画面からメンバーをRealmに保存
