+++ app/src/main/java/com/tokyoc/line_client/Member.kt
@@ -1,10 +1,16 @@
 package com.tokyoc.line_client
 
 import android.os.Parcelable
+import io.realm.RealmObject
+import io.realm.annotations.PrimaryKey
 import kotlinx.android.parcel.Parcelize
 
-// MemberデータFormat
+//MemberデータFormat
 @Parcelize
-data class Member(val name: String,
-                  val id: Int,
-                  val groupId: Int): Parcelable
+open class Member: RealmObject(), Parcelable {
+    @PrimaryKey
+    var id: Int = 0
+    var name: String = "Aさん"
+    var userId: Int = 0
+    var groupId: Int = 0
+}
