+++ app/src/main/res/layout/activity_friend_add.xml
@@ -6,6 +6,7 @@
     android:orientation="vertical">
 
     <EditText
+        android:id="@+id/user_id"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:hint="IDを入力してください" />
